TwentyOneGame
line 28 in twentyOneGame.cs sets the boolean value as false everytime not sure why. the methods i used to instantiate a new list is different than the method used in the video (see Game.cs properties) i wonder if thats the issue, however when i try the same method as the video it whites out the private property and tells me to do it that way. (video on page 100 c#)